Script updates and adds.

+++ activate.sh
@@ -3,17 +3,18 @@ set -e
 OS=$(uname)
 PLATFORM=$(uname -m)
 
-rm -rf dev3.11
-python3.11 -m venv dev3.11
-
-. dev3.11/bin/activate
-
-# Perform the following step only on MacOS, to enable git completions.
-if [ "${OS}" == "Darwin" ]; then
-source ~/.bashrc
+if [ -d dev3.11 ]; then
+  . dev3.11/bin/activate
+else
+  python3.11 -m venv dev3.11
+  . dev3.11/bin/activate
 fi
 
 pip install --upgrade pip
 
-set +e
-pip3 install -r requirements.txt
+if test -f requirements.txt; then
+  echo "File exists."
+  set +e
+  pip3 install -r requirements.txt
+fi
+
